+++ src/services/alignment/alignment.interface.ts
@@ -1,6 +1,0 @@
-/**
- * Allows absolute alignment (AKA positioning) of one element to another
- */
-export interface Alignment
-{
-}