+++ src/services/position/position.utils.ts
@@ -1,0 +1,10 @@
+import {PositionResult} from './position.interface';
+
+/**
+ * Applies `PositionResult` to target element
+ * @param result - Result of positioning process to be applied
+ */
+export function applyPositionResult(result: PositionResult): void
+{
+
+}